Fixed animation bug when falling by setting a boolean
Fixed player running through fall idle loop after a single animation run by adding a setbool of IsJumping to true when in the process of falling. Needs to be set to true here, otherwise the IsJumping boolean is false by default (since it's not set to true through jumping) so it skips the falling idle loop.